% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepfor.R
\name{stepfor}
\alias{stepfor}
\title{stepfor.Rd}
\usage{
stepfor(y = y, d = d, alpha = 0.05)
}
\arguments{
\item{Takes}{a step-wise forward linear regression}

\item{Categorical}{variables are split into columns and assigned value 1 and 0 elsewhere}
}
\description{
stepfor.Rd
}
